// moon-front-end-app/src/app/models/sa-company.model.ts
export class SaCompany {

    companyNo: number;
    companyId: string;
    companyName: string;
    companyNameBangla: string;
    companyAlias: string;
    companyAliasBangla: string;
    companySlogan: string;
    companyAddr1: string;
    companyAddr2: string;
    companyAdd1_Bangla: string;
    companyAdd2_Bangla: string;
    companyCountry: string;
    companyMobile: string;
    officeTime: string;
    companyPhone: string;
    office1Address: string;
    office1TelephoneNo: string;
    office1MobileNo: string;
    office1Email: string;
    office1Webside: string;
    office2Address: string;
    office2TelephoneNo: string;
    office2MobileNo: string;
    office2Email: string;
    office2Webside: string;
    office3Address: string;
    office3TelephoneNo: string;
    office3MobileNo: string;
    office3Email: string;
    office3Webside: string;
    office1AddrName: string;
    office2AddrName: string;
    office3AddrName: string;
    companyFax: string;
    companyEmail: string;
    companyWebsite: string;
    companyLogo: string;
    companyLogoName: string;
    companyFaviconName: string;
    companyDeveloper: string;
    repFooter1: string;
    repFooter2: string;
    repFooterDev: string;
    tinNo: string;
    vatRegNo: string;
    ercNo: string;
    ercDate: Date;
    licenseNo: string;
    companyLogoWeburl: string;
    companyLatestNews: string;
    groupFlag: number;
    activeStat: number;
    ssCreator: number;
    ssCreatedOn: Date;
    ssModifier: number;
    ssModifiedOn: Date;

    constructor(options: any = {}) {
        this.companyNo = options.companyNo || 0;
        this.companyId = options.companyId || '';
        this.companyName = options.companyName || '';
        this.companyNameBangla = options.companyNameBangla || '';
        this.companyAlias = options.companyAlias || '';
        this.companyAliasBangla = options.companyAliasBangla || '';
        this.companySlogan = options.companySlogan || '';
        this.companyAddr1 = options.companyAddr1 || '';
        this.companyAddr2 = options.companyAddr2 || '';
        this.companyAdd1_Bangla = options.companyAdd1_Bangla || '';
        this.companyLatestNews = options.companyLatestNews || '';
        this.companyAdd2_Bangla = options.companyAdd2_Bangla || '';
        this.companyCountry = options.companyCountry || '';
        this.companyMobile = options.companyMobile || '';
        this.officeTime = options.officeTime || '';
        this.companyPhone = options.companyPhone || '';
        this.office1Address = options.office1Address || '';
        this.office1TelephoneNo = options.office1TelephoneNo || '';
        this.office1MobileNo = options.office1MobileNo || '';
        this.office1Email = options.office1Email || '';
        this.office1Webside = options.office1Webside || '';
        this.office2Address = options.office2Address || '';
        this.office2TelephoneNo = options.office2TelephoneNo || '';
        this.office2MobileNo = options.office2MobileNo || '';
        this.office2Email = options.office2Email || '';
        this.office2Webside = options.office2Webside || '';
        this.office3Address = options.office3Address || '';
        this.office3TelephoneNo = options.office3TelephoneNo || '';
        this.office3MobileNo = options.office3MobileNo || '';
        this.office3Email = options.office3Email || '';
        this.office3Webside = options.office3Webside || '';
        this.companyFax = options.companyFax || '';
        this.companyEmail = options.companyEmail || '';
        this.companyWebsite = options.companyWebsite || '';
        this.companyLogo = options.companyLogo || '';
        this.companyLogoName = options.companyLogoName || '';
        this.companyFaviconName = options.companyFaviconName || '';
        this.companyDeveloper = options.companyDeveloper || '';
        this.repFooter1 = options.repFooter1 || '';
        this.repFooter2 = options.repFooter2 || '';
        this.repFooterDev = options.repFooterDev || '';
        this.tinNo = options.tinNo || '';
        this.vatRegNo = options.vatRegNo || '';
        this.ercNo = options.ercNo || '';
        this.ercDate = options.ercDate ? new Date(options.ercDate) : null;
        this.licenseNo = options.licenseNo || '';
        this.office1AddrName = options.office1AddrName || '';
        this.office2AddrName = options.office2AddrName || '';
        this.office3AddrName = options.office3AddrName || '';
        this.companyLogoWeburl = options.companyLogoWeburl || '';
        this.groupFlag = options.groupFlag || null;
        this.activeStat = options.activeStat || null;
        this.ssCreator = options.ssCreator || null;
        this.ssCreatedOn = options.ssCreatedOn ? new Date(options.ssCreatedOn) : null;
        this.ssModifier = options.ssModifier || null;
        this.ssModifiedOn = options.ssModifiedOn ? new Date(options.ssModifiedOn) : null;
       
    }
}